pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zapry-agents-sdk"
version = "1.0.0"
description = "Python SDK for building AI Agents on the Zapry platform with Tool Calling, Memory, Handoff, Guardrails and Tracing"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [
    { name = "CyberFlow Tech", email = "dev@cyberflow.social" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "python-telegram-bot[webhooks]>=20.7",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.urls]
Homepage = "https://github.com/cyberFlowTech/zapry-agents-sdk-python"
Repository = "https://github.com/cyberFlowTech/zapry-agents-sdk-python"

[tool.setuptools.packages.find]
include = ["zapry_agents_sdk*"]
